renamed attacks.rs to square_control to be more descriptive
