+++ rust-implementation/rust-engine/src/moves/mod.rs
@@ -1,4 +1,4 @@
-pub mod attacks;
+pub mod square_control;
 pub mod king;
 pub mod knight;
 pub mod magic;
